Tests/ProvinceTest.cs
using Microsoft.VisualStudio.TestTools.UnitTesting;

[TestClass]
public class ProvinceTest
{
    [TestMethod]
    public void TestTickAndGather()
    {
        Province p = new Province();
        p.Tick();

        Assert.AreEqual(10, p.Gather(ResourceType.Food));
    }

    [TestMethod]
    public void TestProvinceWithCityGeneratesWeapons()
    {
        Province p = new Province(new City("Urbana", 5), new Player());
        p.Tick();
        Assert.AreEqual(5, p.Gather(ResourceType.Weapons));
    }

    [TestMethod]
    public void TestGatherResetsCounter()
    {
        Province p = new Province();
        p.Tick();
        Assert.AreEqual(10, p.Gather(ResourceType.Food));
        Assert.AreEqual(0, p.Gather(ResourceType.Food));
    }
}
